move TryGetSemanticVersion out of ReferenceName
